refactor(cli): extract snapshot subcommands into commands/snapshot.rs
Move run_snapshot + run_snapshot_capture (~985 lines, the biggest handlers)
out of main.rs into commands/snapshot.rs. Minimal coupling — only the EXIT_*
consts + arg structs.

main.rs 5699 -> 4713. Pure move — build + clippy clean.
