+++ crates/cli/src/commands/mod.rs
@@ -1,3 +1,4 @@
 //! CLI subcommand handlers, split out of `main.rs`.
 
 pub mod asset;
+pub mod snapshot;
